requirements.txt
# Deep Learning & ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
sentence-transformers>=2.2.2
datasets>=2.14.0

# Computer Vision & OCR
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3

# Vector DB & Search
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support
pymilvus>=2.3.0

# Table Processing
pandas>=2.0.0
openpyxl>=3.1.0
camelot-py[cv]>=0.11.0
tabula-py>=2.8.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# API & LLM
openai>=1.0.0
anthropic>=0.7.0
tiktoken>=0.5.0
groq
tenacity

# Experiment Tracking
wandb>=0.15.0
mlflow>=2.8.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
click>=8.1.0
pyyaml>=6.0.1
jsonlines>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

# Hugging Face specific models
accelerate>=0.24.0
bitsandbytes>=0.41.0
timm
sentencepiece

# Additional utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0

streamlit